# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Api Documentation",
        "description": " Api Documentation",
        "contact": {
            "name": "Deni Akbar",
            "email": "dnxdenny@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/category_products": {
            "get": {
                "tags": [
                    "CategoryProduct"
                ],
                "summary": "Get list of CategoryProduct",
                "description": "Get list of CategoryProduct",
                "operationId": "getAllCategoryProduct",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CategoryProduct"
                ],
                "summary": "Store CategoryProduct ",
                "description": "Store CategoryProduct ",
                "operationId": "storeCategoryProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/category_products/{id}": {
            "get": {
                "tags": [
                    "CategoryProduct"
                ],
                "summary": "Get CategoryProduct Detail",
                "description": "Get CategoryProduct Detail",
                "operationId": "showCategoryProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of CategoryProduct",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "CategoryProduct"
                ],
                "summary": "Update CategoryProduct",
                "description": "Update CategoryProduct",
                "operationId": "updateCategoryProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of CategoryProduct",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CategoryProduct"
                ],
                "summary": "Delete existing product",
                "description": "",
                "operationId": "deleteCategoryProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}